+++ src/handlers/projects/delete.js
@@ -3,7 +3,7 @@
 const console = require('console');
 const resgen = require('../../lib/resgen');
 const storage = require('../../lib/storage');
-const checkApiKey = require('../../lib/check_api_key');
+const checkApiKey = require('../../lib/checkApiKey');
 const {
     bucketName,
     errorByMessageTable,
