+++ src/handlers/util.js
@@ -4,7 +4,7 @@ const console = require('console');
 const moment = require('moment');
 const resgen = require('../lib/resgen');
 const storage = require('../lib/storage');
-const checkApiKey = require('../lib/check_api_key');
+const checkApiKey = require('../lib/checkApiKey');
 const aws = require('../lib/aws')();
 const kms = aws.kms;
 const lambda = aws.lambda;
