+++ src/lib/checkApiKey.js
