+++ src/lib/messageBuilder.js
